+++ Business/Abstract/IWalletService.cs
@@ -11,8 +11,8 @@ namespace Business.Abstract
         IResult Add(Wallet wallet);
         IResult Update(Wallet wallet);
 
-        IResult BalanceAdd(string userAddress,int amount);
-        IResult BalanceReduction(string userAddress, int amount);
+        IResult BalanceAdd(string userAddress, double amount);
+        IResult BalanceReduction(string userAddress, double amount);
     }
 }
